Created home page and rearranged where buttons and links will take the user
