+++ EcommAssignment2/CSS/contactPageStyle.css
@@ -1,10 +1,4 @@
-﻿body {
-    background: url(../Images/kami_house.jpg) no-repeat center center fixed;
-    background-size: cover;
-    height: auto;
-    font-family: Impact;
-}
-
+﻿
 .body-section {
     margin: auto;
     width: 50%;
