+++ EcommAssignment2/Web.config
@@ -8,11 +8,14 @@
 	<system.webServer>
 		<defaultDocument>
 			<files>
+				<add value="AboutPage.aspx" />
 				<add value="HomePage.aspx" />
 				<add value="SignInPage.aspx" />
 				<add value="MainMenu.aspx" />
 				<add value="ProfilePage.aspx" />
 			    <add value="ContactPage.aspx" />
+				<add value="LoginPage.aspx" />
+			    <add value="orders.aspx" />
 			</files>
 		</defaultDocument>
 		<directoryBrowse enabled="true" />
